+++ src/DocumentFormat.OpenXml/Wordprocessing/TabStop.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using DocumentFormat.OpenXml.Framework.Metadata;
+using DocumentFormat.OpenXml.Framework;
 using System;
 using System.Collections.Generic;
